requirements.txt
# ============================================================
# Install everything with:  pip install -r requirements.txt
# Recommended: use Python 3.10+
# ============================================================

# ── Twitter API ─────────────────────────────────────────────
tweepy==4.14.0              # Official Twitter Python client (supports v2 streams)

# ── Kafka ───────────────────────────────────────────────────
kafka-python==2.0.2         # Low-level Kafka producer/consumer
faust-streaming==0.10.14    # Python stream processing (reads Kafka, processes, writes back)

# ── NLP & ML ────────────────────────────────────────────────
transformers==4.40.0        # Hugging Face: loads RoBERTa model
torch==2.2.2                # PyTorch: required by transformers
vaderSentiment==3.3.2       # Fallback rule-based sentiment (no GPU needed)
gensim==4.3.2               # LDA topic modeling
nltk==3.8.1                 # Tokenization, stopwords

# ── Database ────────────────────────────────────────────────
pymongo==4.7.1              # MongoDB driver
motor==3.4.0                # Async MongoDB driver (used in FastAPI)
redis==5.0.4                # Redis client

# ── API ─────────────────────────────────────────────────────
fastapi==0.111.0            # Web framework with WebSocket support
uvicorn[standard]==0.29.0   # ASGI server to run FastAPI
websockets==12.0            # WebSocket support

# ── Dashboard ───────────────────────────────────────────────
streamlit==1.35.0           # Interactive dashboard
plotly==5.22.0              # Charts: pie, line, bar, scatter
wordcloud==1.9.3            # Word cloud generator
matplotlib==3.9.0           # WordCloud rendering backend
pandas==2.2.2               # DataFrames

# ── Utilities ───────────────────────────────────────────────
python-dotenv==1.0.1        # Load .env file into environment variables
httpx==0.27.0               # Async HTTP requests (used by alert engine)
pydantic==2.7.1             # Data validation (used by FastAPI models)
python-dateutil==2.9.0      # Date parsing
loguru==0.7.2               # Beautiful structured logging
pydantic-settings==2.2.1
